fix(check): warn on invalid <!-- color --> directives
Surface rejected per-slide colour values in kova --check instead of
silently dropping them.

Closes #199
